+++ CHANGELOG.md
@@ -1,5 +1,39 @@
 # Changelog
 
+## [1.24.1](https://github.com/ForkbombEu/didroom-components/compare/v1.24.0...v1.24.1) (2024-07-08)
+
+
+### Bug Fixes
+
+* hidden tabs ([#112](https://github.com/ForkbombEu/didroom-components/issues/112)) ([d40eb8e](https://github.com/ForkbombEu/didroom-components/commit/d40eb8ef2b0e7ccb56ef0ed36ee3adbb6485ae2b))
+
+## [1.24.0](https://github.com/ForkbombEu/didroom-components/compare/v1.23.0...v1.24.0) (2024-07-03)
+
+
+### Features
+
+* tab page layout ([#110](https://github.com/ForkbombEu/didroom-components/issues/110)) ([4d64c85](https://github.com/ForkbombEu/didroom-components/commit/4d64c85c567a640e34cb42ae3b0349732553a686))
+
+
+### Bug Fixes
+
+* center empty state text ([#108](https://github.com/ForkbombEu/didroom-components/issues/108)) ([f65f8e3](https://github.com/ForkbombEu/didroom-components/commit/f65f8e38746563faac533da60e251051cc4c5ca0))
+* header should knows which back method invoke ([#109](https://github.com/ForkbombEu/didroom-components/issues/109)) ([e470ee7](https://github.com/ForkbombEu/didroom-components/commit/e470ee773aab8bb5a03b24ea99aa221dc6eab11d))
+
+## [1.23.0](https://github.com/ForkbombEu/didroom-components/compare/v1.22.0...v1.23.0) (2024-07-01)
+
+
+### Features
+
+* empty state component ([#107](https://github.com/ForkbombEu/didroom-components/issues/107)) ([8887bc2](https://github.com/ForkbombEu/didroom-components/commit/8887bc203571f9230523f6db851872d75860487f))
+* scan button component ([#106](https://github.com/ForkbombEu/didroom-components/issues/106)) ([200983b](https://github.com/ForkbombEu/didroom-components/commit/200983b252284dc8f5c98b431ecc473edc68b3dd))
+
+
+### Bug Fixes
+
+* activity card tests ([#102](https://github.com/ForkbombEu/didroom-components/issues/102)) ([f6e3106](https://github.com/ForkbombEu/didroom-components/commit/f6e3106baca89bd32e0eae657e4e99fad08793e7))
+* header with drawer menu ([#105](https://github.com/ForkbombEu/didroom-components/issues/105)) ([a3f16d3](https://github.com/ForkbombEu/didroom-components/commit/a3f16d349f47601c34d896c9f4a74648753c8fbd))
+
 ## [1.22.0](https://github.com/ForkbombEu/didroom-components/compare/v1.21.0...v1.22.0) (2024-06-19)